Allow the camera to be pointed up and down.
Entirely separate the camera direction from the player direction. Move
the player based on camera direction. The player doesn't now have a
facing direction; if we change it to 3psn the player probably will need
to be rendered facing a particular way, but will probably only change
direction when it moves.

Separate moving from direction entirely. Use WASD for moving and IJKL
for changing direction. This matches the WiiU Zelda games, except for
still being 1psn. This means we need to be able to strafe as well as
walk.