Make a table of keybindings.
I could have put lambdas in the table but I decided it was better to use
string commands, since this is meant to be going in a config file later
on. The command names are currently looked up in globals(); ideally this
should use a separate module so the command names are separated from the
function names in the program.

I may need type annotations of some sort later, since annoyingly Python
won't autoconvert a string to a float. I don't think I can get away with
assuming all commands only take float arguments.