Use numpy for matrices to be passed to GL.
This stores them packed and (I presume) passes them directly as a
double*.

I was hoping I could entirely manipulate the camera angles as
quaternions, but I'm not sure I can. I would need three separate quats
(the initial view correction, the walk direction, the look direction)
and I don't think it's possible to store them multiplied-up as I need to
manipulate them individually. I was hoping I could end up with a single
quat which just gets converted to a matrix and fed to GL but if I can't
it's easier to stick to using angles.