Experiments with the select buffer.
I was getting weird results with things not showing up; I think it was
because I was still using a perspective view and everything drawn ended
up with a window z coord very near 1. This seemed to stop things showing
up when they should.

Also, I seem to need to turn off CULL_FACE, because sometimes when a
culled face and a non-culled face are rendered with the same name stack
neither shows up in the buffer. I suspect this is a bug, but perhaps
not.