Draw the player as a wireframe sphere.
Adjust the camera offset so we can see properly. This perhaps wants to
be a quaternion + scalar offset rather than a fixed vector. The camera
position now changes when we pan, so flag the camera as needing update
rather than the player. (I could do with some sort of signal system
here.)

Having moved the starting point of the camera back we need to increase
the viewing volume or half the platforms don't appear. Hopefully even a
rather large increase at the far end will have less effect on the window z
resolution than a tiny increase at the near end.

The player's ref point is now in the centre, so make the player bump the
floor sooner so we don't fall through.

Fix a bug in the new find_floor_below, even though it's about to be
removed.