Perform selects for bump testing.
Draw GL_POINTs where we would bump.

This appears to work, with caveats. Looking downward works very well;
looking sideways is more difficult, because it's difficult to ensure we
don't get the ground in the view as well, which gives false hits and
seems to mask the real hits. It is necessary to raise the angle of view
enough that the ground isn't visible. (Interestingly selection seems to
be able to see the ground even when the miniview doesn't show anything.)