Clip the selection volume.
Selection is not working properly because the corners of the selection
volume keep clipping into the floor. Attempt to fix this by clipping the
select volume into an octagonal prism; unfortunately, something about this
is causing OpenGL to create segfaults. 3 clip planes seems to be OK; 4
does not work. The clipping is probably not the cause, though, since
rendering with 4 clip planes works just fine.

In any case, selection as a concept appears to be deprecated and is removed
in GL3, so it's entirely possible this is a bug somewhere in the depths of
Intel's GL drivers. I have not investigated thoroughly.