Remove all the selection code.
It doesn't work, and it isn't going to, by the looks of it. I'm just going
to have to either do the physics myself by hand or use Bullet or some
other packaged physics library.

Leave the miniview logic in display.py, since that's potentially useful.
Remove the key binding, since it's now useless.