Start testing text rendering methods.
I was intending to draw text (and a whole 2D overlay) to a
pygame.Surface and then blit that to GL using DrawPixels. Annoyingly I
can't seem to get Surface to give me an RGBA array of pixels. I can get
RGB or A separately but for a blit to GL I would need RGBA.

Use the approach in the tutorial for now: render each glyph to a
texture, and then draw quads. This has the advantage of allowing full 3D
positioning, and the disadvantage of making pixel-accurate rendering
impossible.

I am now relying on (I think) GL 2.0, for non-power-of-2-sized textures.