Implement a pause button.
I have converted the eventloop module into an app class. The transition
is a little awkward as none of the other modules have instance state yet
so the app module needs to store its singleton instance in a global.

Pause brings up a text overlay. Stopping the physics in the app is
decoupled from pause, for things like end-of-level displays.