Make the camera a proper object.
I have moved the walk direction back into the Player; I think this is
wrong, in the end, since the player wants a 'facing' direction which
doesn't change until he walks. But it's easier for now, since the camera
need a reference to the player anyway, for positioning.

input_init now uses bound methods in the dispatch table, which are an
interesting concept. In this situation they just-about substitute for
the lack of closures...