Additions to the world.
Now we can get down to the new platform, make sure we can get back up.
It's not entirely easy, given the need to long-jump. On the bottom
platform underneath the floor would be a good place to put a prize.

Make it so walls can be double-sided. This is tricky: it looks weird if
the walls disappear altogether when we are 'outside', but making them
solid really gets in the way. I think I need some much cleverer camera
positioning logic that won't move through walls; but that needs proper
collision detection.

I tried making the walls semi-transparent but alpha blending and depth
testing are tricky to get right. Possibly this would be easier with GL3?