Make a Texture class which can load from a file.
This was a pig; it seems to be really difficult to get any of these
imaging modules to give you a pointer to a buffer full of bits. This
uses PIL, which works OK but is probably slow, because it copies the
whole thing at least once (maybe twice for the flip).

I make no attempt to check the image mode matches the texture format.
This wouldn't be too hard but I can't be bothered for now.

I am still bothered by what to to about all this rebinding. Maybe I need
a BindManager class under GL which keeps track of what is bound? But
that would be slow, too...