Start working on lighting.
It's tricky to get these objects right: unless we keep rebinding all the
time, it's rather easy to call (say) set_uniform on one object and end up
setting on completely the wrong shader. Maybe trying to make the wrapper
too clean is a mistake; leaky abstractions are a pain.

The set_uniform[34]f methods didn't work; gl..3fv has an extra
'count' parameter which doesn't match gl..1f. And I can't *value,
because it's not always an iterable. Grr.