Use textures of the right pixel depth.
There's no need to use an RGBA texture for one channel.

I was trying to get the specular information into the same texture in
the alpha channel, but I can't get Photoshop to save the image
correctly. It smashes the colour information when it knows the picture
is fully transparent. If it makes a difference I'm sure I could build it
in PIL, or just find/write a PSD loader and pull out whatever channels I
like.