Redo the Shader object.
Dynamically generate a new class for each shader program with methods
for all the uniforms and attribs.

I have not included the *f versions; I think it's easier just to use glm
types everywhere. I have also not yet done anything in particular with
textures; I think the unit->tex mapping wants to be held in the Shader,
not the VAO, so the shader2D setter method wants extending.

I can't decide whether the shader belongs in the VAO object or not yet.