Let Texture determine the image format.
This is not as flexible as it could be; no distinction is made between
wanting a greyscale image to be GL_ALPHA or GL_LUMINANCE, for instance.
But I rather get the impression that GL3 drops all those distinctions in
favour of just using GL_RED for single-channel images (and, presumably,
padding it to RGBA?).