Create a model and import it into my app.
I'm not sure what to do about models which have simple colours vs.
models which have textures, and managing the shader uniforms. It is
tempting to create a big all-in-one shader but I suspect this is not the
best way to go; however I really need some sort of #include for shaders
otherwise I will end up with the Phong code everywhere.