Tidy up the shaders a bit.
The recommendation seems to be not to attempt to link multiple shader
objects of the same type, but to stick to one of each. So I don't need
lists, but I probably do want to implement #include myself.

Stick to the standard extensions, and rename the shaders by what they
do. Delete the extraneous ones (v-light can go because once dead code
has been thrown out v-box/plain.vert is the same).