Draw the keys as actual keys.
The key is much too big, and I have not set up any lighting in the
shader so it renders entirely black, but it draws in the right place.

Camera no longer has a .render method. Instead it just returns a matrix,
and I have a utility function to push that down to FFP when I need to.