Set the light in the shader.
Coordinating the number and types of lights in the world and the shader
is going to get annoying rather quickly, I think.

The key is currently not staying in the same place as the player moves
around. I think I'm using the same matrices everywhere, but I need to
check.