Move the level definition into a file.
This is still raw Python passed to eval(), which I don't like much. But
JSON doesn't allow comments or trailing commas, and YAML is awkward for
writing long lists, and so on... Maybe later.

I am keeping ["floors"] out of the level definition for now, since I
haven't finished the collision stuff yet.