Make all the shaders bind the attribs the same.
Now I don't need to pass a shader in to load a model, and the model
can be used against any shader. (The uniforms have to be looked up
per-shader anyway.)

This means I will have to extend the list if any have extra attribs,
though the attrib methods are still generated if necessary.