Convert Player to glm.
I have left the .walking vector as a vec3 for now; properly it wants to
be a vec2. Perhaps I don't want to use a quat for the rotation at all?
But there's no easy function to apply a 2D rotation to a vec2, unless I
just use a mat2.

I need to stop specifying things in degrees and just use radians
throughout.

We need a fudge factor when putting the player back on top of the floor,
as otherwise the float/double comparison doesn't work right. (This all
needs to go anyway.)