Don't stop dead when we collide.
Keep moving in the plane we collided with.

This doesn't interact with the 'falling' logic yet, so when we land on a
flat surface we still think we are falling. I'm not sure the best way to
fix this; perhaps just check for a sufficiently-vertical normal?

Move general geometry functions, including everything imported from glm,
into .geometry.