Use the new collision logic.
Put the floors in the collision list. I have had to move the starting
position up, as there is a tendency to fall through the floor otherwise;
I'm not entirely sure why. I hope it isn't possible to fall right
through a whole floor in one frame.

We don't move on to the next level yet, as we don't keep track of which
floor we landed on.