Keep checking for hits.
When we slide along the wall, it changes our velocity vector, so we need
to keep checking if the new vector hits anything. A hit always makes the
velocity smaller, so the process will always terminate.

When we reach a corner where we can't move, the cross product goes to
zero. Don't try to normalize that, it makes a mess.