Work out if we are on a corner.
If we are outside the plane but inside the bump margin on more than one
plane, we are at a corner. What needs to happen here is we need to work
out our perpendicular distance to the intersection of all the planes we
are outside, but I haven't worked out how to do that yet.