Draw a key in the overlay.
It was a right pig to get this to be visible.

Despite my reservations about the quality of the text rendering, I don't
think having a pixel-scaled overlay projection is going to work. Dealing
with the display not always being the same aspect ratio is going to be
slightly tricky, too.