Add walls from the plan to the collision list.
Now we have a method to add a block to the list this is easy.

Interestingly, even with every little block implemented as a separate
polyhedron and without any optimisation of the list search, this still
isn't slow. Maybe these levels just aren't big enough to cause a problem
yet? The bump logic doesn't seem to entirely like the not-joined-up
walls, though.