Adjust the models.
There's no need to scale or rotate in the code, the model can be
exported correctly in the first place.

When exporting a .obj from Blender the axis options need to be set
correctly or Blender changes all the coordinates round. (This was
confusing me.) If we don't set Z+-up, Y+-forward in the export box the
coordinates don't come out the same as they were in Blender.