Factor out model code from Items.
Both Key and Portal need to load a model and draw that model, so it's
better to have another layer of superclass that knows how to do that.
The bottom-level classes then just override the little bits they need to
change, and we don't have the same bits of code copied all over the
place.