Start converting text rendering to GL3.
This is a little awkward because we need to keep updating the VBO for
every letter. It would be good to render several letters at once, but I
think the only way to do that in general would be to have all the
letters in one big texture, otherwise we can't rebind the texture
between letters. That then causes potential problems with linear
shading.

And, of course, with GL3 rendering there's a whole lot of state we have
to track instead of letting GL do it.