Give Items an update() method.
I think I want to rename all the physics() methods to update(); they may
in general do things other than physics.

Build the model matrix for items that have one in advance, when
possible, and keep it in a variable. Keep the inverse and normal matrix
too; keeping the inverse will make collision detection easier (we can do
it in model space instead of world space).

Make all the physics() methods take a ctx parameter rather than just a
dt, as Key.update needs now() as well. For the moment this context is
just the Clock object, but that may need to change. (Or Clock may need
to expand into some sort of UpdateContext.)

Items no longer inherit the translate-to-.pos logic; I can't decide
whether this is better or worse.